+++ deploy/5_deploy_prime_factoring_bounty_3072_key.ts
@@ -1,0 +1,4 @@
+import { deployPrimeFactoringBountyFactory } from './utilities/deploy_bounty_factory'
+
+module.exports = deployPrimeFactoringBountyFactory(128, 'PrimeFactoringBountyWithRsaUfo3072Key')
+module.exports.tags = ['PrimeFactoringBounty3072Key']
