+++ deploy/6_deploy_prime_factoring_bounty_4608_key.ts
@@ -1,0 +1,4 @@
+import { deployPrimeFactoringBountyFactory } from './utilities/deploy_bounty_factory'
+
+module.exports = deployPrimeFactoringBountyFactory(192, 'PrimeFactoringBountyWithRsaUfo4608Key')
+module.exports.tags = ['PrimeFactoringBounty4608Key']
