+++ deploy/7_deploy_order_finding_bounty_3072_key.ts
@@ -1,0 +1,4 @@
+import { deployOrderFindingBountyFactory } from './utilities/deploy_bounty_factory'
+
+module.exports = deployOrderFindingBountyFactory(128, 'OrderFindingBountyWithLockGeneration3072Key')
+module.exports.tags = ['OrderFindingBounty3072Key']
