+++ deploy/7_deploy_order_finding_bounty_4608_key.ts
@@ -1,0 +1,4 @@
+import { deployOrderFindingBountyFactory } from './utilities/deploy_bounty_factory'
+
+module.exports = deployOrderFindingBountyFactory(192, 'OrderFindingBountyWithLockGeneration4608Key')
+module.exports.tags = ['OrderFindingBounty4608Key']
